// web/src/api/generated/models/CollectionDetail.ts
export interface CollectionDetail {
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof CollectionDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof CollectionDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof CollectionDetail
     */
    createdBy: string;
    /**
     * 
     * @type {AuthorSummary}
     * @memberof CollectionDetail
     */
    creator: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof CollectionDetail
     */
    id: string;
    /**
     * 
     * @type {Array<CollectionItemDetail>}
     * @memberof CollectionDetail
     */
    items: Array<CollectionItemDetail>;
    /**
     * 
     * @type {string}
     * @memberof CollectionDetail
     */
    spaceId: string;
    /**
     * 
     * @type {string}
     * @memberof CollectionDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof CollectionDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof CollectionDetail
     */
    version: number;
}